rename db-api/citizen.js to user.js
