+++ lib/db-api/index.js
@@ -1,8 +1,8 @@
 /**
- * Expose citizen's database api
+ * Expose user's database api
  */
 
-exports.citizen = require('./citizen');
+exports.user = require('./user');
 
 /**
  * Expose proposal's database api
